🖌️layout : 검색 옵션 컨텐츠들을 별도의 뷰로 분리 및 scrollView 제거
